+++ tests/cli/bootsys/test_platform.py
@@ -30,6 +30,7 @@ import threading
 import unittest
 from unittest import mock
 
+from sat.cli.bootsys.ceph import CephHealthCheckError
 from sat.cli.bootsys.platform import (
     CONTAINER_STOP_SCRIPT,
     ContainerStopThread,
@@ -791,19 +792,18 @@ class TestDoCephFreeze(unittest.TestCase):
         """Set up mocks."""
         # NCN groups are not used by this step of stopping platform services
         self.ncn_groups = {}
-        self.ceph_healthy = mock.patch('sat.cli.bootsys.platform.ceph_healthy').start()
-        self.ceph_healthy.return_value = True
+        self.check_ceph_health = mock.patch('sat.cli.bootsys.platform.check_ceph_health').start()
         self.toggle_ceph_freeze_flags = mock.patch('sat.cli.bootsys.platform.toggle_ceph_freeze_flags').start()
 
     def test_do_ceph_freeze_success(self):
         """Test do_ceph_freeze in the successful case."""
         do_ceph_freeze(self.ncn_groups)
-        self.ceph_healthy.assert_called_once_with(expecting_osdmap_flags=False)
+        self.check_ceph_health.assert_called_once_with(expecting_osdmap_flags=False)
         self.toggle_ceph_freeze_flags.assert_called_once_with(freeze=True)
 
     def test_do_ceph_freeze_unhealthy(self):
         """When Ceph is not healthy, do not freeze Ceph."""
-        self.ceph_healthy.return_value = False
+        self.check_ceph_health.side_effect = CephHealthCheckError
         with self.assertRaises(FatalPlatformError):
             do_ceph_freeze(self.ncn_groups)
         self.toggle_ceph_freeze_flags.assert_not_called()
@@ -816,19 +816,18 @@ class TestDoCephUnfreeze(unittest.TestCase):
         """Set up mocks."""
         # NCN groups are not used by this step of starting platform services
         self.ncn_groups = {}
-        self.ceph_healthy = mock.patch('sat.cli.bootsys.platform.ceph_healthy').start()
-        self.ceph_healthy.return_value = True
+        self.check_ceph_health = mock.patch('sat.cli.bootsys.platform.check_ceph_health').start()
         self.toggle_ceph_freeze_flags = mock.patch('sat.cli.bootsys.platform.toggle_ceph_freeze_flags').start()
 
     def test_do_ceph_unfreeze_success(self):
         """Test do_ceph_unfreeze in the successful case."""
         do_ceph_unfreeze(self.ncn_groups)
-        self.ceph_healthy.assert_called_once_with(expecting_osdmap_flags=True)
+        self.check_ceph_health.assert_called_once_with(expecting_osdmap_flags=True)
         self.toggle_ceph_freeze_flags.assert_called_once_with(freeze=False)
 
     def test_do_ceph_unfreeze_unhealthy(self):
         """When Ceph is not healthy, do not unfreeze Ceph."""
-        self.ceph_healthy.return_value = False
+        self.check_ceph_health.side_effect = CephHealthCheckError
         with self.assertRaises(FatalPlatformError):
             do_ceph_unfreeze(self.ncn_groups)
         self.toggle_ceph_freeze_flags.assert_not_called()